Add calendar view, make the User controller RESTful, design, miscellaneous fixes

+++ app/views/readCourses.blade.php
@@ -1,46 +1,62 @@
 @extends('_master')
 
 @section('title')
-Quizical | Courses
+Quizical | Your Courses
 @stop
 
 @section('content')
-	<h1>Your courses</h1>
+
+	@if ($user->id == Auth::user()->id)
+		<h1>Your courses</h1>
+	@else
+		<h1>{{ $user->first_name }} {{ $user->last_name }}'s Courses</h1>
+	@endif
 
 	{{-- List courses --}}
 
 	<div class="row">
+		@if(count($courses) == 0)
+			@if ($user->id == Auth::user()->id)
+				<p>You don't have any courses!</p>
+			@else
+				<p>This user doesn't have any courses.</p>
+			@endif
+		@endif
+
 		@foreach($courses as $course)
 			<div class="col-md-12">
 				<div class="panel panel-default">
 					<div class="panel-body">
 						<h3 class="courseBlock"><a href="/courses/{{$course->id}}">{{ $course->name }}</a> <br> 
-						<small>{{ $course->block }} Block, {{ $course->start_year }}-{{ $course->start_year+1 }}</small></h3>
+						<small>Block {{ $course->block }}, {{ $course->start_year }}-{{ $course->start_year+1 }}</small></h3>
 
-						<div class="col-md-1 crudButtons">
-							<div class="col-md-6">
-								{{-- Edit button --}}
-								{{ Form::open(array('route' => array('courses.edit', $course->id), 'method' => 'get')) }}
-									<button class="btn btn-warning btn-sm" href="/courses/{{$course->id}}/edit">Edit</button>
-								{{ Form::close() }}
-							</div>
+						@if ($user->id == Auth::user()->id)
+							<div class="col-md-1 crudButtons">
+								<div class="col-md-6">
+									{{-- Edit button --}}
+									{{ Form::open(array('route' => array('courses.edit', $course->id), 'method' => 'get')) }}
+										<button class="btn btn-warning btn-sm" href="/courses/{{$course->id}}/edit">Edit</button>
+									{{ Form::close() }}
+								</div>
 
-							<div class="col-md-6">
-								{{-- Delete button --}}
-								{{ Form::open(array('route' => array('courses.destroy', $course->id), 'method' => 'delete')) }}
-									<button type="submit" class="btn btn-danger btn-sm" href="{{ URL::route('courses.destroy', $course->id) }}">Delete</button>
-								{{ Form::close() }}
+								<div class="col-md-6">
+									{{-- Delete button --}}
+									{{ Form::open(array('route' => array('courses.destroy', $course->id), 'method' => 'delete')) }}
+										<button type="submit" class="btn btn-danger btn-sm" href="{{ URL::route('courses.destroy', $course->id) }}">Delete</button>
+									{{ Form::close() }}
+								</div>
 							</div>
-						</div>
+						@endif
 					</div>
 				</div>
 			</div>
 		@endforeach
 	</div>
 
-	{{-- Add button --}}
-
-	{{ Form::open(array('route' => 'courses.create', 'method' => 'get')) }}
-		<button class="btn btn-primary" href="/courses/create">Add a course</button>
-	{{ Form::close() }}
+	@if ($user->id == Auth::user()->id)
+		{{-- Add button --}}
+		{{ Form::open(array('route' => 'courses.create', 'method' => 'get')) }}
+			<button class="btn btn-primary" href="/courses/create">Add a course</button>
+		{{ Form::close() }}
+	@endif
 @stop
